test(discord): split group policy coverage out of the oversized channel test file
